requirements.txt
easyocr
pillow
ollama
colorama
pyautogui
pyperclip
requests